+++ app/models/__init__.py
@@ -1,0 +1,2 @@
+from .event_model import *
+from .verification_model import *
