+++ event_management/settings.py
@@ -44,6 +44,7 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     "rest_framework",
     "rest_framework_simplejwt",
+    'app',
 ]
 
 MIDDLEWARE = [
